Move AugurError from utils to new file errors.py
This reduces complexity and chance of circular imports in utils.
