+++ augur/filter.py
@@ -18,9 +18,10 @@ from tempfile import NamedTemporaryFile
 from typing import Collection
 
 from .dates import numeric_date, numeric_date_type, SUPPORTED_DATE_HELP_TEXT, is_date_ambiguous, get_numerical_dates
+from .errors import AugurError
 from .index import index_sequences, index_vcf
 from .io import open_file, read_metadata, read_sequences, write_sequences, is_vcf as filename_is_vcf, write_vcf
-from .utils import AugurError, read_strains
+from .utils import read_strains
 
 comment_char = '#'
