+++ augur/parse.py
@@ -6,7 +6,7 @@ import sys
 
 from .io import open_file, read_sequences, write_sequences
 from .dates import get_numerical_date_from_value
-from .utils import AugurError
+from .errors import AugurError
 
 forbidden_characters = str.maketrans(
     {' ': None,
